Move PeerPool to peer.py
